# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Customer Controller
  description: ""
  termsOfService: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Joel Silva
    email: joeltadeu@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
  - url: http://localhost:9081
    description: Generated server url
...
paths:
...
  /v1/customers/{customerId}/documents/{id}:
    get:
      tags:
        - Document Controller
      summary: Retrieve a customer document by id
      description: This endpoint is responsible to retrieve the customer document
        data by id
      operationId: findDocumentById
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: Id of the document to be searched
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      responses:
        "404":
          description: Document not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to retrieve
                    a document does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Document with id 7 was not found
                    date: 2022-10-28T21:39:54.603263862
        "200":
          description: Return document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        "500":
          description: An unexpected error occurred during retrieve the document
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to retrieve a customer document
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
    put:
      tags:
        - Document Controller
      summary: Update a document
      description: This endpoint is responsible to update a customer document data
        by id
      operationId: updateDocument
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: Id of the document to be updated
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentRequest'
        required: true
      responses:
        "200":
          description: Document updated
          content:
            application/json: {}
        "404":
          description: Document not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to update
                    a document does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Document with id 7 was not found
                    date: 2022-10-28T21:39:54.603263862
        "500":
          description: An unexpected error occurred during update the document
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to update a document
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "400":
          description: Document request is invalid
          content:
            application/json:
              examples:
                Bad request example:
                  description: A bad request response example when trying to update
                    a document
                  value: |
                    {
                         "code": 400,
                         "status": "Bad Request",
                         "description": "Validation Exception",
                         "date": "2022-11-25T20:14:04.078313400",
                         "attributes": [
                             {
                                 "attribute": "type",
                                 "errors": [
                                     "Document Type cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "documentNumber",
                                 "errors": [
                                     Document number cannot be null"
                                 ]
                             }
                         ]
                    }
    delete:
      tags:
        - Document Controller
      summary: Delete the customer document by id
      description: This endpoint is responsible to delete the customer document by
        id
      operationId: deleteDocument
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: Id of the document to be deleted
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      responses:
        "204":
          description: Document deleted
          content:
            application/json: {}
        "500":
          description: An unexpected error occurred during delete the customer document
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to delete customer document
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "404":
          description: Document not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to delete
                    a customer document does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Document with id 7 was not found
                    date: 2022-10-28T21:39:54.603263862
...
  /v1/customers/{customerId}/documents:
    get:
      tags:
        - Document Controller
      summary: Retrieve a list of documents by customer id
      description: This endpoint is responsible to retrieve a list of customer documents
        based on the applied filters
      operationId: listAllDocuments
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: filter
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/DocumentFilter'
      responses:
        "200":
          description: Return customer documents list
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageDocumentResponse'
        "500":
          description: An unexpected error occurred during retrieve the customer documents
            list
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to retrieve a customer documents list
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
    post:
      tags:
        - Document Controller
      summary: Create a document
      description: This endpoint is responsible to create a new document
      operationId: createDocument
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentRequest'
        required: true
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        "400":
          description: Document request is invalid
          content:
            application/json:
              examples:
                Bad request example:
                  description: A bad request response example when trying to create
                    a document
                  value: |
                    {
                         "code": 400,
                         "status": "Bad Request",
                         "description": "Validation Exception",
                         "date": "2022-11-25T20:14:04.078313400",
                         "attributes": [
                             {
                                 "attribute": "type",
                                 "errors": [
                                     "Document Type cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "documentNumber",
                                 "errors": [
                                     Document number cannot be null"
                                 ]
                             }
                         ]
                    }
        "500":
          description: An unexpected error occurred during creating the document
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to create a document
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
...
components:
  schemas:
...
    DocumentRequest:
      required:
        - documentNumber
        - type
      type: object
      properties:
        type:
          type: string
          description: Document type
          example: PASSPORT
          enum:
            - DRIVER_LICENSE
            - PASSPORT
            - PPS
        documentNumber:
          type: string
          description: Document number
          example: 187263IN
...
    DocumentResponse:
      required:
        - documentNumber
      type: object
      properties:
        id:
          type: integer
          description: Document Id
          format: int64
          example: 13
        type:
          type: string
          description: Document type
          example: PASSPORT
          enum:
            - DRIVER_LICENSE
            - PASSPORT
            - PPS
        documentNumber:
          type: string
          description: Document number
          example: 187263IN
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
...
    DocumentFilter:
      required:
        - pageNumber
        - pageSize
      type: object
      properties:
        pageNumber:
          type: integer
          description: Zero-base page index (0..N)
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: Number of records per page
          format: int32
          example: 10
          default: 10
        sortDirection:
          type: string
          description: Sort direction
          example: ASC|DESC
          enum:
            - ASC
            - DESC
        sortBy:
          type: string
          description: Sort column name
          example: firstName
        type:
          type: string
          description: Document type to be searched
          example: PPS|DRIVER_LICENSE|PASSPORT
          enum:
            - DRIVER_LICENSE
            - PASSPORT
            - PPS
        documentNumber:
          type: string
          description: Document number to be searched
          example: 09817387IH
    PageDocumentResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
...